Enable AutoExtern to consume from multiple source files (#2292)
